Initial commit
This function is designed to trigger when an image is uploaded to the rawimages container. It will resize the image such that it has a width of 500 whilst the height is determined dynamically. The aspect ratio of the image will be maintained. The image will then be saved in the processedimages container.